+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 23.5.1
+NEW ENHANCEMENTS:
+* resoruce/cvo_aws: add `assume_role_arn` for AWS CVO HA.
+
+## 23.5.0
+BUG FIXES:
+* resource/cbs: fix cbs to work without volume.
+
 ## 23.5.0
 BUG FIXES:
 * resource/connector_gcp: fix label key and value do not acccpt only numeric value.